Refactor lib to services.
