+++ pages/index.tsx
@@ -1,8 +1,8 @@
 import Head from 'next/head'
-import Layout from '../components/Layout';
-import config from '../configs/blogConfig';
-import {paginate} from '@/lib/Paginator';
-import {getAllPosts} from "@/lib/Posts";
+import Layout from '@/components/Layout';
+import config from '@/configs/blogConfig';
+import {paginate} from '@/services/PaginatorService';
+import {getAllPosts} from "@/services/PostService";
 import PaginatedPostsList from "@/components/PaginatedPostsList";
 import {DEFAULT_ITEMS_PER_PAGE} from "@/constants/pagination";
 import PaginationControls from "@/components/PaginationControls";
