+++ tsconfig.json
@@ -18,7 +18,7 @@
       "@/components/*": ["components/*"],
       "@/configs/*": ["configs/*"],
       "@/constants/*": ["constants/*"],
-      "@/lib/*": ["lib/*"],
+      "@/services/*": ["services/*"],
       "@/types/*": ["types/*"],
     }
   },
